+++ src/main/java/roomescape/controller/ReservationController.java
@@ -10,7 +10,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseBody;
-import roomescape.dao.TimeDao;
 import roomescape.dto.ReservationRequestDto;
 import roomescape.domain_entity.Reservation;
 import roomescape.service.ReservationService;
@@ -20,8 +19,6 @@ public class ReservationController {
 
     @Autowired
     private ReservationService reservationService;
-    @Autowired
-    private TimeDao timeDao;
 
     @GetMapping("/admin")
     public String displayMain() {
