+++ src/test/java/roomescape/controller/ReservationControllerTest.java
@@ -24,91 +24,9 @@ import roomescape.domain_entity.ReservationTime;
 import roomescape.dto.TimeRequestDto;
 import roomescape.service.TimeService;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 public class ReservationControllerTest {
 
-    @Autowired
-    private JdbcTemplate jdbcTemplate;
-    @Autowired
-    private TimeService timeService;
-    @Autowired
-    private ReservationController reservationController;
-
-    @Test
-    @DisplayName("예약 관리 메인 페이지를 렌더링한다.")
-    void displayMainAdmin() {
-        RestAssured.given().log().all()
-                .when().get("/admin")
-                .then().log().all()
-                .statusCode(200);
-    }
-
-    @Test
-    @DisplayName("예약을 생성한다.")
-    void createReservation() {
-        //given
-        timeService.createTime(new TimeRequestDto(LocalTime.of(10, 0, 0)));
-        Map<String, Object> reservation = new HashMap<>();
-        reservation.put("name", "브라운");
-        reservation.put("date", "2023-08-05");
-        reservation.put("timeId", 1);
-
-        //when
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(reservation)
-                .when().post("/reservations")
-                .then().log().all()
-                .statusCode(200)
-                .body("id", is(1),
-                        "name", is("브라운"),
-                        "date", is("2023-08-05"),
-                        "time.id", is(1),
-                        "time.startAt", is("10:00:00")
-                );
-
-        //then
-        RestAssured.given().log().all()
-                .when().get("/reservations")
-                .then().log().all()
-                .statusCode(200)
-                .body("size()", is(1));
-    }
-
-    @Test
-    @DisplayName("전체 예약을 조회한다.")
-    void readAllReservations() {
-        createReservation();
-
-        RestAssured.given().log().all()
-                .when().get("/admin/reservation")
-                .then().log().all()
-                .statusCode(200);
-
-        RestAssured.given().log().all()
-                .when().get("/reservations")
-                .then().log().all()
-                .statusCode(200)
-                .body("size()", is(1));
-    }
-
-    @Test
-    @DisplayName("예약을 삭제한다.")
-    void deleteReservation() {
-        createReservation();
-
-        RestAssured.given().log().all()
-                .when().delete("/reservations/1")
-                .then().log().all()
-                .statusCode(200);
-
-        RestAssured.given().log().all()
-                .when().get("/reservations")
-                .then().log().all()
-                .statusCode(200)
-                .body("size()", is(0));
-    }
+    private final ReservationController reservationController = new ReservationController();
 
     @Test
     @DisplayName("컨트롤러는 DB 접근 책임을 가지지 않는다.")
